+++ pkg/bindinfo/BUILD.bazel
@@ -5,15 +5,12 @@ go_library(
     srcs = [
         "binding.go",
         "binding_cache.go",
-        "binding_match.go",
         "global_handle.go",
         "session_handle.go",
     ],
     importpath = "github.com/pingcap/tidb/pkg/bindinfo",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/bindinfo/internal/logutil",
-        "//pkg/bindinfo/norm",
         "//pkg/kv",
         "//pkg/metrics",
         "//pkg/parser",
@@ -31,6 +28,7 @@ go_library(
         "//pkg/util/chunk",
         "//pkg/util/hack",
         "//pkg/util/hint",
+        "//pkg/util/logutil",
         "//pkg/util/parser",
         "//pkg/util/sqlexec",
         "@com_github_dgraph_io_ristretto//:ristretto",
@@ -46,32 +44,26 @@ go_test(
     timeout = "moderate",
     srcs = [
         "binding_cache_test.go",
-        "binding_match_test.go",
-        "cross_db_binding_test.go",
         "global_handle_test.go",
         "main_test.go",
-        "optimize_test.go",
         "session_handle_test.go",
     ],
     embed = [":bindinfo"],
     flaky = True,
     race = "on",
-    shard_count = 34,
+    shard_count = 31,
     deps = [
-        "//pkg/bindinfo/internal",
-        "//pkg/bindinfo/norm",
         "//pkg/parser",
         "//pkg/parser/ast",
         "//pkg/parser/auth",
         "//pkg/parser/format",
-        "//pkg/parser/mysql",
+        "//pkg/parser/terror",
         "//pkg/server",
         "//pkg/session/types",
         "//pkg/sessionctx/variable",
         "//pkg/testkit",
         "//pkg/testkit/testsetup",
-        "//pkg/types",
-        "//pkg/util",
+        "//pkg/util/parser",
         "//pkg/util/stmtsummary",
         "@com_github_ngaut_pools//:pools",
         "@com_github_pingcap_failpoint//:failpoint",
