+++ pkg/bindinfo/binding_cache_test.go
@@ -15,11 +15,13 @@
 package bindinfo
 
 import (
+	"strings"
 	"testing"
 	"time"
 
-	"github.com/pingcap/tidb/pkg/bindinfo/norm"
 	"github.com/pingcap/tidb/pkg/parser"
+	"github.com/pingcap/tidb/pkg/parser/ast"
+	"github.com/pingcap/tidb/pkg/parser/format"
 	"github.com/pingcap/tidb/pkg/sessionctx/variable"
 	"github.com/stretchr/testify/require"
 )
@@ -28,7 +30,7 @@ func bindingNoDBDigest(t *testing.T, b *Binding) string {
 	p := parser.New()
 	stmt, err := p.ParseOneStmt(b.BindSQL, b.Charset, b.Collation)
 	require.NoError(t, err)
-	_, noDBDigest := norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
+	_, noDBDigest := NormalizeStmtForBinding(stmt, WithoutDB(true))
 	return noDBDigest
 }
 
@@ -100,4 +102,55 @@ func TestBindCache(t *testing.T) {
 		}
 		return hit == 2
 	}, time.Second*5, time.Millisecond*100)
+}
+
+func getTableName(n []*ast.TableName) []string {
+	result := make([]string, 0, len(n))
+	for _, v := range n {
+		var sb strings.Builder
+		restoreFlags := format.RestoreKeyWordLowercase
+		restoreCtx := format.NewRestoreCtx(restoreFlags, &sb)
+		v.Restore(restoreCtx)
+		result = append(result, sb.String())
+	}
+	return result
+}
+
+func TestExtractTableName(t *testing.T) {
+	tc := []struct {
+		sql    string
+		tables []string
+	}{
+		{
+			"select /*+ HASH_JOIN(t1, t2) */ * from t1 t1 join t1 t2 on t1.a=t2.a where t1.b is not null;",
+			[]string{"t1", "t1"},
+		},
+		{
+			"select * from t",
+			[]string{"t"},
+		},
+		{
+			"select * from t1, t2, t3;",
+			[]string{"t1", "t2", "t3"},
+		},
+		{
+			"select * from t1 where t1.a > (select max(a) from t2);",
+			[]string{"t1", "t2"},
+		},
+		{
+			"select * from t1 where t1.a > (select max(a) from t2 where t2.a > (select max(a) from t3));",
+			[]string{"t1", "t2", "t3"},
+		},
+		{
+			"select a,b,c,d,* from t1 where t1.a > (select max(a) from t2 where t2.a > (select max(a) from t3));",
+			[]string{"t1", "t2", "t3"},
+		},
+	}
+	for _, tt := range tc {
+		stmt, err := parser.New().ParseOneStmt(tt.sql, "", "")
+		require.NoError(t, err)
+		rs := CollectTableNames(stmt)
+		result := getTableName(rs)
+		require.Equal(t, tt.tables, result)
+	}
 }
