--- pkg/bindinfo/internal/BUILD.bazel
@@ -1,15 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "internal",
-    srcs = ["testutil.go"],
-    importpath = "github.com/pingcap/tidb/pkg/bindinfo/internal",
-    visibility = ["//pkg/bindinfo:__subpackages__"],
-    deps = [
-        "//pkg/parser",
-        "//pkg/parser/ast",
-        "//pkg/testkit",
-        "//pkg/util/parser",
-        "@com_github_stretchr_testify//require",
-    ],
-)
