--- pkg/bindinfo/norm/BUILD.bazel
@@ -1,13 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "norm",
-    srcs = ["normalize.go"],
-    importpath = "github.com/pingcap/tidb/pkg/bindinfo/norm",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/parser",
-        "//pkg/parser/ast",
-        "//pkg/util/parser",
-    ],
-)
