+++ pkg/bindinfo/tests/BUILD.bazel
@@ -5,25 +5,27 @@ go_test(
     timeout = "moderate",
     srcs = [
         "bind_test.go",
+        "cross_db_binding_test.go",
         "main_test.go",
     ],
     flaky = True,
     race = "on",
-    shard_count = 18,
+    shard_count = 21,
     deps = [
         "//pkg/bindinfo",
-        "//pkg/bindinfo/internal",
-        "//pkg/bindinfo/norm",
         "//pkg/domain",
         "//pkg/meta/model",
         "//pkg/parser",
         "//pkg/parser/model",
-        "//pkg/parser/terror",
+        "//pkg/parser/mysql",
+        "//pkg/session/types",
         "//pkg/testkit",
         "//pkg/testkit/testsetup",
+        "//pkg/types",
         "//pkg/util",
         "//pkg/util/parser",
         "//pkg/util/stmtsummary",
+        "@com_github_ngaut_pools//:pools",
         "@com_github_stretchr_testify//require",
         "@org_uber_go_goleak//:goleak",
     ],
