+++ pkg/bindinfo/tests/bind_test.go
@@ -23,19 +23,24 @@ import (
 	"testing"
 
 	"github.com/pingcap/tidb/pkg/bindinfo"
-	"github.com/pingcap/tidb/pkg/bindinfo/internal"
-	"github.com/pingcap/tidb/pkg/bindinfo/norm"
 	"github.com/pingcap/tidb/pkg/domain"
 	"github.com/pingcap/tidb/pkg/meta/model"
 	"github.com/pingcap/tidb/pkg/parser"
 	pmodel "github.com/pingcap/tidb/pkg/parser/model"
-	"github.com/pingcap/tidb/pkg/parser/terror"
 	"github.com/pingcap/tidb/pkg/testkit"
 	"github.com/pingcap/tidb/pkg/util"
 	utilparser "github.com/pingcap/tidb/pkg/util/parser"
 	"github.com/pingcap/tidb/pkg/util/stmtsummary"
 	"github.com/stretchr/testify/require"
 )
+
+// utilCleanBindingEnv cleans the binding environment.
+func utilCleanBindingEnv(tk *testkit.TestKit) {
+	tk.MustExec("update mysql.bind_info set status='deleted' where source != 'builtin'")
+	tk.MustExec(`admin reload bindings`)
+	tk.MustExec("delete from mysql.bind_info where source != 'builtin'")
+	tk.MustExec(`admin reload bindings`)
+}
 
 func TestPrepareCacheWithBinding(t *testing.T) {
 	store := testkit.CreateMockStore(t)
@@ -275,130 +280,6 @@ func TestPrepareCacheWithBinding(t *testing.T) {
 	tk.MustUseIndexForConnection(strconv.FormatUint(tkProcess.ID, 10), "ib(b)")
 }
 
-// TestBindingSymbolList tests sql with "?, ?, ?, ?", fixes #13871
-func TestBindingSymbolList(t *testing.T) {
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, b int, INDEX ia (a), INDEX ib (b));")
-	tk.MustExec("insert into t value(1, 1);")
-
-	// before binding
-	tk.MustQuery("select a, b from t where a = 3 limit 1, 100")
-	require.Equal(t, "t:ia", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
-	tk.MustUseIndex("select a, b from t where a = 3 limit 1, 100", "ia(a)")
-
-	tk.MustExec(`create global binding for select a, b from t where a = 1 limit 0, 1 using select a, b from t use index (ib) where a = 1 limit 0, 1`)
-
-	// after binding
-	tk.MustQuery("select a, b from t where a = 3 limit 1, 100")
-	require.Equal(t, "t:ib", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
-	tk.MustUseIndex("select a, b from t where a = 3 limit 1, 100", "ib(b)")
-
-	// Normalize
-	stmt, err := parser.New().ParseOneStmt("select a, b from test . t where a = 1 limit 0, 1", "", "")
-	require.NoError(t, err)
-
-	_, noDBDigest := norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
-	binding, matched := dom.BindHandle().MatchGlobalBinding(tk.Session(), noDBDigest, bindinfo.CollectTableNames(stmt))
-	require.True(t, matched)
-	require.Equal(t, "select `a` , `b` from `test` . `t` where `a` = ? limit ...", binding.OriginalSQL)
-	require.Equal(t, "SELECT `a`,`b` FROM `test`.`t` USE INDEX (`ib`) WHERE `a` = 1 LIMIT 0,1", binding.BindSQL)
-	require.Equal(t, "test", binding.Db)
-	require.Equal(t, bindinfo.Enabled, binding.Status)
-	require.NotNil(t, binding.Charset)
-	require.NotNil(t, binding.Collation)
-	require.NotNil(t, binding.CreateTime)
-	require.NotNil(t, binding.UpdateTime)
-}
-
-// TestBindingInListWithSingleLiteral tests sql with "IN (Lit)", fixes #44298
-func TestBindingInListWithSingleLiteral(t *testing.T) {
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, b int, INDEX ia (a), INDEX ib (b));")
-	tk.MustExec("insert into t value(1, 1);")
-
-	// GIVEN
-	sqlcmd := "select a, b from t where a in (1)"
-	bindingStmt := `create global binding for select a, b from t where a in (1, 2, 3) using select a, b from t use index (ib) where a in (1, 2, 3)`
-
-	// before binding
-	tk.MustQuery(sqlcmd)
-	require.Equal(t, "t:ia", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
-	tk.MustUseIndex(sqlcmd, "ia(a)")
-
-	tk.MustExec(bindingStmt)
-
-	// after binding
-	tk.MustQuery(sqlcmd)
-	require.Equal(t, "t:ib", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
-	tk.MustUseIndex(sqlcmd, "ib(b)")
-
-	tk.MustQuery("select @@last_plan_from_binding").Check(testkit.Rows("1"))
-
-	// Normalize
-	stmt, err := parser.New().ParseOneStmt("select a, b from test . t where a in (1)", "", "")
-	require.NoError(t, err)
-
-	_, noDBDigest := norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
-	binding, matched := dom.BindHandle().MatchGlobalBinding(tk.Session(), noDBDigest, bindinfo.CollectTableNames(stmt))
-	require.True(t, matched)
-	require.Equal(t, "select `a` , `b` from `test` . `t` where `a` in ( ... )", binding.OriginalSQL)
-	require.Equal(t, "SELECT `a`,`b` FROM `test`.`t` USE INDEX (`ib`) WHERE `a` IN (1,2,3)", binding.BindSQL)
-	require.Equal(t, "test", binding.Db)
-	require.Equal(t, bindinfo.Enabled, binding.Status)
-	require.NotNil(t, binding.Charset)
-	require.NotNil(t, binding.Collation)
-	require.NotNil(t, binding.CreateTime)
-	require.NotNil(t, binding.UpdateTime)
-}
-
-func TestBestPlanInBaselines(t *testing.T) {
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, b int, INDEX ia (a), INDEX ib (b));")
-	tk.MustExec("insert into t value(1, 1);")
-
-	// before binding
-	tk.MustQuery("select a, b from t where a = 3 limit 1, 100")
-	require.Equal(t, "t:ia", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
-	tk.MustUseIndex("select a, b from t where a = 3 limit 1, 100", "ia(a)")
-
-	tk.MustQuery("select a, b from t where b = 3 limit 1, 100")
-	require.Equal(t, "t:ib", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
-	tk.MustUseIndex("select a, b from t where b = 3 limit 1, 100", "ib(b)")
-
-	tk.MustExec(`create global binding for select a, b from t where a = 1 limit 0, 1 using select /*+ use_index(@sel_1 test.t ia) */ a, b from t where a = 1 limit 0, 1`)
-	tk.MustExec(`create global binding for select a, b from t where b = 1 limit 0, 1 using select /*+ use_index(@sel_1 test.t ib) */ a, b from t where b = 1 limit 0, 1`)
-
-	stmt, _, _ := internal.UtilNormalizeWithDefaultDB(t, "select a, b from t where a = 1 limit 0, 1")
-
-	_, noDBDigest := norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
-	binding, matched := dom.BindHandle().MatchGlobalBinding(tk.Session(), noDBDigest, bindinfo.CollectTableNames(stmt))
-	require.True(t, matched)
-	require.Equal(t, "select `a` , `b` from `test` . `t` where `a` = ? limit ...", binding.OriginalSQL)
-	require.Equal(t, "SELECT /*+ use_index(@`sel_1` `test`.`t` `ia`)*/ `a`,`b` FROM `test`.`t` WHERE `a` = 1 LIMIT 0,1", binding.BindSQL)
-	require.Equal(t, "test", binding.Db)
-	require.Equal(t, bindinfo.Enabled, binding.Status)
-
-	tk.MustQuery("select a, b from t where a = 3 limit 1, 10")
-	require.Equal(t, "t:ia", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
-	tk.MustUseIndex("select a, b from t where a = 3 limit 1, 100", "ia(a)")
-
-	tk.MustQuery("select a, b from t where b = 3 limit 1, 100")
-	require.Equal(t, "t:ib", tk.Session().GetSessionVars().StmtCtx.IndexNames[0])
-	tk.MustUseIndex("select a, b from t where b = 3 limit 1, 100", "ib(b)")
-}
-
 func TestIssue50646(t *testing.T) {
 	store := testkit.CreateMockStore(t)
 	tk := testkit.NewTestKit(t, store)
@@ -410,40 +291,6 @@ func TestIssue50646(t *testing.T) {
 	tk.MustHavePlan(`delete /*+ inl_merge_join(t) */ t, t1 from t ignore index(idx) join t1 ignore index(idx) on t.a=t1.a;`, "MergeJoin")
 	tk.MustExec(`delete t, t1 from t ignore index(idx) join t1 ignore index(idx) on t.a=t1.a;`)
 	tk.MustQuery(`select @@last_plan_from_binding`).Check(testkit.Rows("1"))
-}
-
-func TestErrorBind(t *testing.T) {
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustGetErrMsg("create global binding for select * from t using select * from t", "[schema:1146]Table 'test.t' doesn't exist")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec("create table t(i int, s varchar(20))")
-	tk.MustExec("create table t1(i int, s varchar(20))")
-	tk.MustExec("create index index_t on t(i,s)")
-
-	_, err := tk.Exec("create global binding for select * from t where i>100 using select * from t use index(index_t) where i>100")
-	require.NoError(t, err, "err %v", err)
-
-	stmt, err := parser.New().ParseOneStmt("select * from test . t where i > ?", "", "")
-	require.NoError(t, err)
-	_, noDBDigest := norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
-	binding, matched := dom.BindHandle().MatchGlobalBinding(tk.Session(), noDBDigest, bindinfo.CollectTableNames(stmt))
-	require.True(t, matched)
-	require.Equal(t, "select * from `test` . `t` where `i` > ?", binding.OriginalSQL)
-	require.Equal(t, "SELECT * FROM `test`.`t` USE INDEX (`index_t`) WHERE `i` > 100", binding.BindSQL)
-	require.Equal(t, "test", binding.Db)
-	require.Equal(t, bindinfo.Enabled, binding.Status)
-	require.NotNil(t, binding.Charset)
-	require.NotNil(t, binding.Collation)
-	require.NotNil(t, binding.CreateTime)
-	require.NotNil(t, binding.UpdateTime)
-
-	tk.MustExec("drop index index_t on t")
-	require.Equal(t, 1, len(tk.MustQuery(`show global bindings`).Rows()))
-	tk.MustQuery("select * from t where i > 10")
 }
 
 func TestStmtHints(t *testing.T) {
@@ -463,65 +310,6 @@ func TestStmtHints(t *testing.T) {
 	require.Equal(t, uint64(0), tk.Session().GetSessionVars().StmtCtx.MaxExecutionTime)
 	// TODO(crazycs520): Fix me.
 	//require.Equal(t, uint64(0), tk.Session().GetSessionVars().GetTiKVClientReadTimeout())
-}
-
-func TestHintsSetID(t *testing.T) {
-	store, dom := testkit.CreateMockStoreAndDomain(t)
-
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int, index idx_a(a))")
-	tk.MustExec("create global binding for select * from t where a > 10 using select /*+ use_index(test.t, idx_a) */ * from t where a > 10")
-	// Verify the added Binding contains ID with restored query block.
-	stmt, err := parser.New().ParseOneStmt("select * from t where a > ?", "", "")
-	require.NoError(t, err)
-	_, noDBDigest := norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
-	binding, matched := dom.BindHandle().MatchGlobalBinding(tk.Session(), noDBDigest, bindinfo.CollectTableNames(stmt))
-	require.True(t, matched)
-	require.Equal(t, "select * from `test` . `t` where `a` > ?", binding.OriginalSQL)
-	require.Equal(t, "use_index(@`sel_1` `test`.`t` `idx_a`)", binding.ID)
-
-	internal.UtilCleanBindingEnv(tk)
-	tk.MustExec("create global binding for select * from t where a > 10 using select /*+ use_index(t, idx_a) */ * from t where a > 10")
-	_, noDBDigest = norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
-	binding, matched = dom.BindHandle().MatchGlobalBinding(tk.Session(), noDBDigest, bindinfo.CollectTableNames(stmt))
-	require.True(t, matched)
-	require.Equal(t, "select * from `test` . `t` where `a` > ?", binding.OriginalSQL)
-	require.Equal(t, "use_index(@`sel_1` `test`.`t` `idx_a`)", binding.ID)
-
-	internal.UtilCleanBindingEnv(tk)
-	tk.MustExec("create global binding for select * from t where a > 10 using select /*+ use_index(@sel_1 t, idx_a) */ * from t where a > 10")
-	_, noDBDigest = norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
-	binding, matched = dom.BindHandle().MatchGlobalBinding(tk.Session(), noDBDigest, bindinfo.CollectTableNames(stmt))
-	require.True(t, matched)
-	require.Equal(t, "select * from `test` . `t` where `a` > ?", binding.OriginalSQL)
-	require.Equal(t, "use_index(@`sel_1` `test`.`t` `idx_a`)", binding.ID)
-
-	internal.UtilCleanBindingEnv(tk)
-	tk.MustExec("create global binding for select * from t where a > 10 using select /*+ use_index(@qb1 t, idx_a) qb_name(qb1) */ * from t where a > 10")
-	_, noDBDigest = norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
-	binding, matched = dom.BindHandle().MatchGlobalBinding(tk.Session(), noDBDigest, bindinfo.CollectTableNames(stmt))
-	require.True(t, matched)
-	require.Equal(t, "select * from `test` . `t` where `a` > ?", binding.OriginalSQL)
-	require.Equal(t, "use_index(@`sel_1` `test`.`t` `idx_a`)", binding.ID)
-
-	internal.UtilCleanBindingEnv(tk)
-	tk.MustExec("create global binding for select * from t where a > 10 using select /*+ use_index(T, IDX_A) */ * from t where a > 10")
-	_, noDBDigest = norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
-	binding, matched = dom.BindHandle().MatchGlobalBinding(tk.Session(), noDBDigest, bindinfo.CollectTableNames(stmt))
-	require.True(t, matched)
-	require.Equal(t, "select * from `test` . `t` where `a` > ?", binding.OriginalSQL)
-	require.Equal(t, "use_index(@`sel_1` `test`.`t` `idx_a`)", binding.ID)
-
-	internal.UtilCleanBindingEnv(tk)
-	err = tk.ExecToErr("create global binding for select * from t using select /*+ non_exist_hint() */ * from t")
-	require.True(t, terror.ErrorEqual(err, parser.ErrParse))
-	tk.MustExec("create global binding for select * from t where a > 10 using select * from t where a > 10")
-	_, noDBDigest = norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
-	binding, matched = dom.BindHandle().MatchGlobalBinding(tk.Session(), noDBDigest, bindinfo.CollectTableNames(stmt))
-	require.True(t, matched)
-	require.Equal(t, "select * from `test` . `t` where `a` > ?", binding.OriginalSQL)
 }
 
 func TestBindingWithIsolationRead(t *testing.T) {
@@ -680,7 +468,7 @@ func TestBindSQLDigest(t *testing.T) {
 	}
 	for _, c := range cases {
 		stmtsummary.StmtSummaryByDigestMap.Clear()
-		internal.UtilCleanBindingEnv(tk)
+		utilCleanBindingEnv(tk)
 		sql := "create global binding for " + c.origin + " using " + c.hint
 		tk.MustExec(sql)
 		res := tk.MustQuery(`show global bindings`).Rows()
@@ -771,7 +559,7 @@ func TestDropBindBySQLDigest(t *testing.T) {
 	h := dom.BindHandle()
 	// global scope
 	for _, c := range cases {
-		internal.UtilCleanBindingEnv(tk)
+		utilCleanBindingEnv(tk)
 		sql := "create global binding for " + c.origin + " using " + c.hint
 		tk.MustExec(sql)
 		h.LoadFromStorageToCache(true)
@@ -788,7 +576,7 @@ func TestDropBindBySQLDigest(t *testing.T) {
 
 	// session scope
 	for _, c := range cases {
-		internal.UtilCleanBindingEnv(tk)
+		utilCleanBindingEnv(tk)
 		sql := "create binding for " + c.origin + " using " + c.hint
 		tk.MustExec(sql)
 		res := tk.MustQuery(`show bindings`).Rows()
@@ -823,27 +611,6 @@ func TestJoinOrderHintWithBinding(t *testing.T) {
 	require.Equal(t, res[0][0], "select * from ( `test` . `t1` join `test` . `t2` on `t1` . `a` = `t2` . `a` ) left join `test` . `t3` on `t2` . `b` = `t3` . `b`")
 
 	tk.MustExec("drop global binding for select * from t1 join t2 on t1.a=t2.a join t3 on t2.b=t3.b")
-}
-
-func TestNormalizeStmtForBinding(t *testing.T) {
-	tests := []struct {
-		sql        string
-		normalized string
-		digest     string
-	}{
-		{"select 1 from b where (x,y) in ((1, 3), ('3', 1))", "select ? from `b` where row ( `x` , `y` ) in ( ... )", "ab6c607d118c24030807f8d1c7c846ec23e3b752fd88ed763bb8e26fbfa56a83"},
-		{"select 1 from b where (x,y) in ((1, 3), ('3', 1), (2, 3))", "select ? from `b` where row ( `x` , `y` ) in ( ... )", "ab6c607d118c24030807f8d1c7c846ec23e3b752fd88ed763bb8e26fbfa56a83"},
-		{"select 1 from b where (x,y) in ((1, 3), ('3', 1), (2, 3),('x', 'y'))", "select ? from `b` where row ( `x` , `y` ) in ( ... )", "ab6c607d118c24030807f8d1c7c846ec23e3b752fd88ed763bb8e26fbfa56a83"},
-		{"select 1 from b where (x,y) in ((1, 3), ('3', 1), (2, 3),('x', 'y'),('x', 'y'))", "select ? from `b` where row ( `x` , `y` ) in ( ... )", "ab6c607d118c24030807f8d1c7c846ec23e3b752fd88ed763bb8e26fbfa56a83"},
-		{"select 1 from b where (x) in ((1), ('3'), (2),('x'),('x'))", "select ? from `b` where ( `x` ) in ( ( ... ) )", "03e6e1eb3d76b69363922ff269284b359ca73351001ba0e82d3221c740a6a14c"},
-		{"select 1 from b where (x) in ((1), ('3'), (2),('x'))", "select ? from `b` where ( `x` ) in ( ( ... ) )", "03e6e1eb3d76b69363922ff269284b359ca73351001ba0e82d3221c740a6a14c"},
-	}
-	for _, test := range tests {
-		stmt, _, _ := internal.UtilNormalizeWithDefaultDB(t, test.sql)
-		n, digest := norm.NormalizeStmtForBinding(stmt, norm.WithoutDB(true))
-		require.Equal(t, test.normalized, n)
-		require.Equal(t, test.digest, digest)
-	}
 }
 
 func showBinding(tk *testkit.TestKit, showStmt string) [][]any {
